requirements.txt
alembic==0.8.7
blinker==1.4
cffi==1.7.0
click==6.6
cryptography==1.5
dominate==2.2.1
enum34==1.1.6
Flask==0.10.1
Flask-Bootstrap==3.3.5.7
Flask-HTTPAuth==3.2.0
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Migrate==1.6.0
Flask-Moment==0.5.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
idna==2.1
ipaddress==1.0.16
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.4
MarkupSafe==0.23
MySQL-python==1.2.3rc1
numpy==1.11.1
openvas-lib==1.1.2
paramiko==2.0.2
pyasn1==0.1.9
pycparser==2.14
python-editor==1.0.1
python-nmap==0.6.0
six==1.10.0
SQLAlchemy==1.0.14
visitor==0.1.3
Werkzeug==0.11.11
whois==0.7
WTForms==2.1
xlrd==1.0.0
XlsxWriter==0.9.3
